requirements.txt
PyYAML==6.0.2
qbittorrent-api==2025.4.1
requests==2.32.3
setuptools==78.1.0
urllib3==2.4.0
fastapi==0.109.2
uvicorn==0.27.1
